# swagger.yaml
openapi: 3.0.0
info:
  title: Trashcan API
  version: 1.0.0
  description: API documentation for our intelligent trashcans
servers:
  - url: http://api.wurstkatze.eu/{version}
    description: Production
    variables:
      version:
        default: v1
        enum:
          - v1
  - url: http://localhost:4000/{version}
    description: Local
    variables:
      version:
        default: v1
        enum:
          - v1
...
paths:
...
  /percentage/{trashcanName}:
    get:
      tags:
        - Trashcans
      summary: Get fill percentage for a specific trashcan
      parameters:
        - name: trashcanName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  percentage:
                    type: number
        '404':
          description: Not Found
  /percentage/{trashcanName}/statistics:
    get:
      tags:
        - Trashcans
      summary: Get statistics for fill percentage of a specific trashcan over time
      parameters:
        - name: trashcanName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                    percentage:
                      type: number
        '404':
          description: Not Found